Add new ExpressionCommand class
Currently it contains only `_view` and `_filters` methods, which were
previously defined in ListCommand. Now both EditCommand and ListCommand
will inherit from it to solve problems inside EditCommand.

This fixes #28
